added logic to form when owner is user
