+++ app/assets/stylesheets/components/_form.scss
@@ -1,3 +1,12 @@
 abbr[title], abbr[data-original-title] {
   text-decoration: none;
 }
+
+.fas.fa-sad-tear {
+  color: #CB6B16;
+}
+
+.tip {
+  font-style: italic;
+  font-weight: lighter;
+}
