+++ app/views/shared/_booking_form.html.erb
@@ -1,8 +1,13 @@
-<div class="col-xs-6 col-xs-offset-3">
-  <%= simple_form_for [@bike, @booking] do |f| %>
-    <%= f.input :start_date, input_html:  { class: "select2 custom__input" } %>
-    <%= f.input :end_date, input_html:  { class: "select2 custom__input" } %>
-    <%= hidden_field_tag(:bike_id, @bike.id) %>
-    <%= f.submit "Book this bike!", id: "booking__submit" %>
-  <% end %>
-</div>
+<% if bike.user != current_user %>
+  <div class="col-xs-6 col-xs-offset-3">
+    <%= simple_form_for [@bike, @booking] do |f| %>
+      <%= f.input :start_date, input_html:  { class: "select2 custom__input" } %>
+      <%= f.input :end_date, input_html:  { class: "select2 custom__input" } %>
+      <%= hidden_field_tag(:bike_id, @bike.id) %>
+      <%= f.submit "Book this bike!", id: "booking__submit" %>
+    <% end %>
+  </div>
+  <% else %>
+  <p>Sorry, you can't book your own bike <i class="fas fa-sad-tear"></i></p>
+  <span class="tip">(psttt, just don't put it up for rent if you want to ride it...)</span>
+<% end %>
